Added Master and User Prefabs
For the Artboard root prefab we now generate 2 Prefabs.
1. A "normal" prefab (MasterPrefab) in the MasterPrefabs folder that has the guid of the symbolInstance as its name
2. A prefab variant (UserPrefab) derived from the Masterprefab that gets stored with a proper name in the UserPrefabs folder

This way the user can work with the UserPrefab an make changes to it without loosing them when the MasterPrefab gets overridden next time the UI gets updated by a Designer
TODO:
Currently this system only can keep changes that are made to the root prefab (ArtBoard). All changes that are made to components (nested prefbs) will still be overriden on the next ui import
The problem here is that unity does not allow us to replace components in a prefab and therefore we cant change the nested prefabs to prefab variants in the UserPrefab